Move features -> system and separate capybara config
